Java refactor complete
