+++ java/compaction/compaction-rust/pom.xml
@@ -25,6 +25,11 @@
     <modelVersion>4.0.0</modelVersion>
     <artifactId>compaction-rust</artifactId>
     <dependencies>
+        <dependency>
+            <groupId>com.github.jnr</groupId>
+            <artifactId>jnr-ffi</artifactId>
+            <version>${jnr.ffi.version}</version>
+        </dependency>
         <!-- Sleeper dependencies -->
         <dependency>
             <groupId>sleeper</groupId>
