+++ java/ffi/pom.xml
@@ -27,12 +27,6 @@
     <artifactId>ffi</artifactId>
 
     <dependencies>
-        <!-- Sleeper dependencies -->
-        <dependency>
-            <groupId>sleeper</groupId>
-            <artifactId>core</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
         <!-- Native code loading and linking -->
         <dependency>
             <groupId>com.github.jnr</groupId>
